Fix k_sparse calculation type and add test result CSV files for TSP algorithms

+++ tsp/test_tsplib_m.py
@@ -39,7 +39,7 @@ def infer_instance(model, pyg_data, distances, n_ants, t_aco_diff, n_runs=1):
             distances.cpu(),
             n_ants,
             heuristic=heu_mat.cpu() if heu_mat is not None else heu_mat,
-            k_sparse=len(distances)/10,
+            k_sparse=int(len(distances)/10),
         )
 
         results = torch.zeros(size=(len(t_aco_diff),))
